Urls optimizadas, Uso de postgres sql, Alojar directorio de apps
