+++ apps/ecommerce/decorators.py
@@ -4,7 +4,7 @@ from django.shortcuts import redirect
 def unauthenticated_user(view_func):
     def wrapper_func(request, *args, **kwargs):
         if request.user.is_authenticated:
-            return redirect("ecommerce:home")
+            return redirect("home")
         else:
             return view_func(request, *args, **kwargs)
