+++ apps/ecommerce/forms.py
@@ -1,6 +1,4 @@
 from django import forms
-from django.contrib.auth.forms import UserCreationForm
-from django.contrib.auth.models import User
 
 from .models import Image, Product
 
@@ -36,8 +34,3 @@ class ImageForm(forms.ModelForm):
         fields = ["image_location"]
 
 
-class UserForm(UserCreationForm):    
-
-    class Meta:
-        model = User
-        fields = ["username", "email", "password1", "password2"]
