+++ apps/users/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from .views import register_view, login_view, logout_view
+
+urlpatterns = [
+    path("register/", register_view, name="register"),
+    path("login/", login_view, name="login"),
+    path("logout/", logout_view, name="logout"),
+]
