+++ templates/base.html
@@ -23,7 +23,7 @@
 
     {% include 'navbar.html' %}
 
-    <section style="min-height: 25rem;" class="m-0 p-0">
+    <section style="min-height: 35rem;" class="m-0 p-0">
         {% block content %}
         {% endblock content %}
     </section>
