+++ templates/navbar.html
@@ -2,7 +2,7 @@
     <div class="container">
         <div class="d-flex flex-wrap align-items-center justify-content-center justify-content-lg-start">
             <a class="d-flex mb-2 mb-lg-0 mx-3 btn btn-outline-light"
-            href="{% url 'ecommerce:home' %}"> AisModa marketPlace</a>
+            href="{% url 'home' %}"> AisModa marketPlace</a>
 
             <form class="col-12 col-lg-auto mb-3 mb-lg-0 me-lg-auto">
                 <input type="search" class="form-control form-control-dark" placeholder="Search..." aria-label="Search">
@@ -16,14 +16,13 @@
 
                 {% if request.user.is_authenticated and request.user.is_staff%}
                 <span class="user-msg h3 mx-3"> Hello, {{request.user}}</span>
-                <a class="btn btn-warning m-1" href="{% url 'ecommerce:logout' %}"> Logout </a>
-                <a class="btn btn-warning m-1" href="{% url 'ecommerce:product-create' %}">new product</a>
+                <a class="btn btn-warning m-1" href="{% url 'logout' %}"> Logout </a>
+                <a class="btn btn-warning m-1" href="{% url 'product-create' %}">new product</a>
                 {% elif request.user.is_authenticated %}
                 <span class="user-msg h3 mx-3"> Hello, {{request.user}}</span>
-                <a class="btn btn-warning m-1" href="{% url 'ecommerce:logout' %}"> Logout </a>
+                <a class="btn btn-warning m-1" href="{% url 'logout' %}"> Logout </a>
                 {% else %}
-                <a class="btn btn-outline-light me-2 m-3" href="{% url 'ecommerce:login' %}"> Login </a>
-                <!-- <a class="btn btn-outline-light me-2 m-3" href="{% url 'ecommerce:register' %}"> register </a> -->
+                <a class="btn btn-outline-light me-2 m-3" href="{% url 'login' %}"> Login </a>
                 {% endif %}
             </div>
         </div>
